Fixed a bunch of problems, everything works together at MVP level. Wooo terrible commit messages!

+++ dir_watcher.py
@@ -1,40 +1,39 @@
-import os
-from watchdog.observers import Observer
 from watchdog.events import FileSystemEventHandler
-import time
 from file_transfer import *
 
-class DirectoryWatcher( FileSystemEventHandler ):
-	def __init__(self, user):
-		FileSystemEventHandler.__init__(self)
-		self.user = user
 
-	def on_created(self, event):
-		if not event.is_directory:
-			file_upload(event.src_path, self.user)
-		else:
-			# Do we need to handle directories?
-			pass
+class DirectoryWatcher(FileSystemEventHandler):
+    def __init__(self, user, flag):
+        FileSystemEventHandler.__init__(self)
+        self.user = user
+        self.flag = flag
 
-	def on_deleted(self, event):
-		if not event.is_directory:
-			file_delete(event.src_path)
-		else:
-			# Do we need to handle directories?
-			pass
-		
+    def on_created(self, event):
+        if not event.is_directory:
+            # Commenting this because it causes double-uploads
+            pass   # file_upload(event.src_path, self.user)
+        else:
+            # Do we need to handle directories?
+            pass
 
-	def on_modified(self, event):
-		if not event.is_directory:
-			file_upload(event.src_path, self.user)
-		else:
-			# Do we need to handle directories?
-			pass
+    def on_deleted(self, event):
+        if not event.is_directory and not self.flag.is_set():
+            file_delete(event.src_path)
+        else:
+            # Do we need to handle directories?
+            pass
 
-	def on_moved(self, event):
-		if not event.is_directory:
-			file_delete(event.src_path)
-			file_upload(event.dest_path, self.user)
-		else:
-			# Do we need to handle directories?
-			pass
+    def on_modified(self, event):
+        if not event.is_directory and not self.flag.is_set():
+            file_upload(event.src_path, self.user)
+        else:
+            # Do we need to handle directories?
+            pass
+
+    def on_moved(self, event):
+        if not event.is_directory:
+            file_delete(event.src_path)
+            file_upload(event.dest_path, self.user)
+        else:
+            # Do we need to handle directories?
+            pass
